// ticketmayo-front/src/components/Tickets/Type.tsx
export interface Idata {
    ticketInfo: {
        id: number;
        title: string;
        open: string;
        openTime: string;
        playLoc: string;
        grade: string;
        img: string;
        price: number;
        cnt: number;
        payNumber: number;
        seatInfo: string;
        deliverFee: number;
    };
}